LayoutTests/http/tests/security/dataURL/xss-DENIED-from-data-url-sub-frame-to-data-url-sub-frame-expected.txt

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL data:text/html,%3Chtml%3E%3Chead%3E%3Cscript%3Efunction%20fireSentinel()%20{if%20(window.layoutTestController)layoutTestController.globalFlag%20=%20true;}%3C/script%3E%3C/head%3E%3Cbody%20onload=%22fireSentinel();%22%3E%3Cp%3EInner%20iframe.%3C/p%3E%3Cp%20id='accessMe'%3EPass:%20Cross%20frame%20access%20from%20a%20sibling%20data:%20URL%20was%20denied.%3C/p%3E%3C/body%3E%3C/html%3E from frame with URL data:text/html,%3Chtml%3E%3Chead%3E%3Cscript%3Efunction%20loaded()%20{if%20(window.layoutTestController)%20{setTimeout(waitForFlag,%201);function%20waitForFlag()%20{if%20(!layoutTestController.globalFlag)%20{setTimeout(waitForFlag,%201);return;}performTest();}}%20else%20{document.getElementById('console').innerHTML%20=%20'Click%20the%20button%20when%20all%20the%20subframes%20have%20finished%20loading.';}}function%20performTest()%20{try%20{parent.frames['firstSubFrame'].document.getElementById('accessMe').innerHTML%20=%20'FAIL:%20Cross%20frame%20access%20from%20a%20sibling%20data:%20URL%20was%20allowed.';}%20catch%20(e)%20{}if%20(window.layoutTestController)layoutTestController.notifyDone();}%3C/script%3E%3C/head%3E%3Cbody%20onload=%22loaded();%22%3E%3Cp%3EInner%20iframe.%3C/p%3E%3Cbutton%20onclick='performTest();'%3ETest%3C/button%3E%3Cpre%20id='console'%3E%3C/pre%3E%3C/body%3E%3C/html%3E. Domains, protocols and ports must match.

This tests that a data: URL subframe can't access a sibling data: URL subframe.

  

--------
Frame: 'firstSubFrame'
--------
Inner iframe.

Pass: Cross frame access from a sibling data: URL was denied.

--------
Frame: 'secondSubFrame'
--------
Inner iframe.

Test
